backend/requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot
sqlalchemy
pydantic
pydantic-settings
python-dotenv
websockets
openai
httpx
